Upgrade library to API version 2019-05-29. (#84)
## Description

Upgrade library to support API version `2019-05-29`

### More information (if any)  

- Removed the flexibility to change API version because this updates
will support only the latest API version `2019-05-29`.
- Updated few fields according to the latest API version.
- Updated tests.

### Documentation (if any)

None for now. We need to update our README to announce the new major
version release that supports only the latest API version `2019` and
guide users on which library version to use for different API version.

## Decision(s)

We removed the ability to set the API version. This change is necessary
because the upcoming updates will not be compatible with the older API
versions. So, switching API version to an older version is impractical.
Instead, we specified the API version directly in the library.
 
## Rollback procedure

<!-- Describe the Rollback procedure. If none/default, update with
`default rollback procedure` -->